requirements.txt
bcrypt==4.3.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1